[Bugfix] Enable API server scale-out using Dockerfile entrypoint
